+++ frontend-web/src/pages/Footer/styles.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  color: #fff;
+  margin-top: 50px;
+  font-weight: bolder;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-content: space-between;
+`;