+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 
+## [0.5.0] - 2022-01-24
+
+### Added
+- Confirmação de remoção de lançamento: adição de modal antes de remover
+- Exportação de dados: cria um json com todos dados cadastrados para download.
+- Iniciar dia: remove todos lançamentos e adiciona o do inicio de dia.
+- Link para wiki no rodapé
+### Changed
+### Removed
+### Fixed
+- banco de horas nao mostrando valor ao abrir tela
+
 ## [0.4.0] - 2021-11-22
 ### Added
 - Banco de horas: função para carregar e salvar minutos do banco de horas criado com horas extras ou devidas de lançamentos anteriores.
@@ -70,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [Unreleased]: https://github.com/alexferreiradev/gerenciador_horas_trabalho/compare/v0.2.0...HEAD
+[0.5.0]: https://github.com/alexferreiradev/gerenciador_horas_trabalho/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/alexferreiradev/gerenciador_horas_trabalho/compare/v0.3.0...v0.4.0
 [0.2.0]: https://github.com/alexferreiradev/gerenciador_horas_trabalho/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/alexferreiradev/gerenciador_horas_trabalho/compare/v0.0.1...v0.1.0